requirements.txt
python-telegram-bot>=20.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-ollama>=0.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytz>=2023.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0